+++ Enlighter.php
@@ -41,6 +41,7 @@ if (version_compare(phpversion(), '5.3', '>=')){
 	require_once(ENLIGHTER_PLUGIN_PATH.'/class/ContextualHelp.php');
 	require_once(ENLIGHTER_PLUGIN_PATH.'/class/ConfigGenerator.php');
 	require_once(ENLIGHTER_PLUGIN_PATH.'/class/ObjectCache.php');
+    require_once(ENLIGHTER_PLUGIN_PATH.'/class/GoogleWebfontResources.php');
 	
 	// run enlighter
 	Enlighter::run();