+++ class/GoogleWebfontResources.php
@@ -1,0 +1,23 @@
+<?php
+// AUTO GENERATED FILE - DO NOT EDIT
+namespace Enlighter;
+class GoogleWebfontResources{
+public static function getMonospaceFonts(){
+return array (
+  'Anonymous Pro' => 'Anonymous+Pro:regular,italic,700,700italic',
+  'Cousine' => 'Cousine:regular,italic,700,700italic',
+  'Cutive Mono' => 'Cutive+Mono:regular',
+  'Droid Sans Mono' => 'Droid+Sans+Mono:regular',
+  'Fira Mono' => 'Fira+Mono:regular,700',
+  'Inconsolata' => 'Inconsolata:regular,700',
+  'Nova Mono' => 'Nova+Mono:regular',
+  'Oxygen Mono' => 'Oxygen+Mono:regular',
+  'PT Mono' => 'PT+Mono:regular',
+  'Roboto Mono' => 'Roboto+Mono:regular,italic,700,700italic',
+  'Share Tech Mono' => 'Share+Tech+Mono:regular',
+  'Source Code Pro' => 'Source+Code+Pro:regular,700',
+  'Ubuntu Mono' => 'Ubuntu+Mono:regular,italic,700,700italic',
+  'VT323' => 'VT323:regular',
+);
+}
+}